fix response datatype

+++ backend/de.metas.shipper.client.nshift/src/main/java/de/metas/shipper/client/nshift/json/response/JsonShipmentResponse.java
@@ -70,7 +70,7 @@ public class JsonShipmentResponse
 	String submitDt;
 
 	@JsonProperty("Vol")
-	Integer vol;
+	Long vol;
 
 	@JsonProperty("Weight")
 	Integer weight;
